% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{extendr_svg}
\alias{extendr_svg}
\title{A graphic device that does nothing}
\usage{
extendr_svg(svg_file, width, height)
}
\arguments{
\item{svg_file}{A path to output SVG file.}

\item{width}{Device width in inch.}

\item{height}{Device width in inch.}
}
\description{
A graphic device that does nothing
}
